feat(notification): Redesign completed jobs alerts
MomentJS Warnings
- fix MomentJS deprecation warnings with .add(). The parameters need to
be switched.

Redux/Configurable polling interval time
- redux was integrated so that the polling interval can be controlled by
 the kappnav-config ConfigMap, specifically the
'ui-completed-jobs-polling-interval-in-ms' property.  An example would
be:
ui-completed-jobs-polling-interval-in-ms: '5000'

Polling behavior changes
- Changed the polling to always make an initial API call on browser load
 or refresh.  Before, a polling would increment the time scope for
completed jobs.  The time scope is no longer based on this idea.  The
time scope for completed jobs is now based on the last seen completed
job.  This means that each poll no longer increments the time scope and
the polling is functionally harmless (except performance).

Timestamps
- OpenShift 4.3 stores timestamps in YYYY-MM-DDTHH:MM:ssZ format.  When
this format is translated into the JSON from the API, the format becomes
 YYYY-MM-HH-ss.SSSZ.  Note the SSS is 000 because OpenShift only stores
the timestamps to the seconds.