fix: Add better API error handling in Notification
- Add catch() blocks to prevent the UI from going blank from a runtime
error.  This blank UI happened in GitHub kappnav/issues#152.

- Add a new method to process HTTP 207 responses.  In the API code, 207
appears when the Kubernetes Java client encounters an error.  The error
is return as a 207 response with a JSON message in the body  In the
method checkForErrorFromCommandActionAPI(), throw was chosen over using
Promise.reject().  throw is used to terminate the code flow.